pyproject.toml
[project]
name = "radar-economico-merida"
version = "0.1.0"
description = "Dashboard geoespacial end-to-end en Streamlit para analizar la actividad economica de Merida con ETL versionado."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "streamlit>=1.43",
    "pandas>=2.2",
    "requests>=2.32",
    "pyarrow>=19.0",
    "pydeck>=0.9",
    "geopandas>=1.0",
    "shapely>=2.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[tool.uv]
package = false
